chore: Rename GitSettings to GitInstances
Closes #624
